Move app side bar to a separate component
