+++ apps/gitness/src/components-v2/app-shell.tsx
@@ -1,27 +1,18 @@
-import { memo, useCallback, useEffect, useMemo, useState } from 'react'
-import { Outlet, useLocation, useNavigate, useParams } from 'react-router-dom'
+import { FC, memo, useEffect, useMemo } from 'react'
+import { Outlet, useParams } from 'react-router-dom'
 
-import {
-  ManageNavigation,
-  MenuGroupType,
-  MenuGroupTypes,
-  MoreSubmenu,
-  NavbarItemType,
-  SettingsMenu,
-  Sidebar
-} from '@harnessio/ui/components'
-import { MainContentLayout, SidebarView } from '@harnessio/ui/views'
+import { NavbarItemType } from '@harnessio/ui/components'
+import { MainContentLayout } from '@harnessio/ui/views'
 
 import { useNav } from '../components/stores/recent-pinned-nav-links.store'
-import { getNavbarMenuData } from '../data/navbar-menu-data'
 import { getPinnedMenuItemsData } from '../data/pinned-menu-items-data'
-import { useAppContext } from '../framework/context/AppContext'
 import { useRoutes } from '../framework/context/NavigationContext'
 import { useLocationChange } from '../framework/hooks/useLocationChange'
 import { useRepoImportEvents } from '../framework/hooks/useRepoImportEvent'
 import { useSelectedSpaceId } from '../framework/hooks/useSelectedSpaceId'
 import { useTranslationStore } from '../i18n/stores/i18n-store'
 import { PathParams } from '../RouteDefinitions'
+import { AppSideBar } from './app-side-bar'
 import Breadcrumbs from './breadcrumbs/breadcrumbs'
 import BreadcrumbsMFE from './breadcrumbs/breadcrumbs-mfe'
 import { Toaster } from './toaster'
@@ -33,20 +24,13 @@ interface NavLinkStorageInterface {
   }
 }
 
-export const AppShell = () => {
+export const AppShell: FC = () => {
   const routes = useRoutes()
-  const { currentUser } = useAppContext()
-  const navigate = useNavigate()
-  const location = useLocation()
-  const { pinnedMenu, recentMenu, setPinned, setRecent, setNavLinks } = useNav()
+  const { spaceId } = useParams<PathParams>()
+  const { pinnedMenu, setRecent, setNavLinks } = useNav()
   const { t } = useTranslationStore()
-  const { spaceId, repoId } = useParams<PathParams>()
   const selectedSpaceId = useSelectedSpaceId(spaceId)
   const spaceIdPathParam = spaceId ?? selectedSpaceId ?? ''
-
-  const [showMoreMenu, setShowMoreMenu] = useState(false)
-  const [showSettingMenu, setShowSettingMenu] = useState(false)
-  const [showCustomNav, setShowCustomNav] = useState(false)
 
   const pinnedMenuItemsData = useMemo(
     () => getPinnedMenuItemsData({ t, routes, spaceId: spaceIdPathParam }),
@@ -77,145 +61,15 @@ export const AppShell = () => {
     }
   }, [spaceIdPathParam])
 
-  /**
-   * Map mock data menu by type to Settings and More
-   */
-  const { moreMenu, settingsMenu } = useMemo(() => {
-    const navbarMenuData = getNavbarMenuData({
-      t,
-      routes,
-      spaceId: spaceIdPathParam,
-      repoId
-    })
-    return navbarMenuData.reduce<{
-      moreMenu: MenuGroupType[]
-      settingsMenu: MenuGroupType[]
-    }>(
-      (acc, item) => {
-        if (item.type === MenuGroupTypes.SETTINGS) {
-          acc.settingsMenu.push(item)
-        } else {
-          acc.moreMenu.push(item)
-        }
-
-        return acc
-      },
-      {
-        moreMenu: [],
-        settingsMenu: []
-      }
-    )
-  }, [t, routes, spaceIdPathParam, repoId])
-
-  /**
-   * Handle logout
-   */
-  const handleLogOut = () => navigate(routes.toLogout())
-
-  /**
-   * Toggle show more menu
-   */
-  const handleMoreMenu = useCallback((state?: boolean) => {
-    setShowSettingMenu(false)
-    setShowMoreMenu(prevState => state ?? !prevState)
-  }, [])
-
-  /**
-   * Toggle system settings menu
-   */
-  const handleSettingsMenu = useCallback((state?: boolean) => {
-    setShowMoreMenu(false)
-    setShowSettingMenu(prevState => state ?? !prevState)
-  }, [])
-
-  /**
-   * Toggle custom navigation modal
-   */
-  const handleCustomNav = useCallback(() => {
-    setShowCustomNav(prevState => !prevState)
-  }, [])
-
-  /**
-   * Close all menu when location changed
-   */
-  useEffect(() => {
-    setShowMoreMenu(false)
-    setShowSettingMenu(false)
-    setShowCustomNav(false)
-  }, [location])
-
-  /**
-   * Handle save recent and pinned items
-   */
-  const handleSave = (nextRecentItems: NavbarItemType[], nextPinnedItems: NavbarItemType[]) => {
-    setNavLinks({
-      pinnedMenu: nextPinnedItems,
-      recentMenu: nextRecentItems
-    })
-  }
-
-  /**
-   * Remove recent menu item
-   */
-  const handleRemoveRecentMenuItem = useCallback(
-    (item: NavbarItemType) => {
-      setRecent(item, true)
-    },
-    [setRecent]
-  )
-
-  /**
-   * Change pinned menu items
-   */
-  const handleChangePinnedMenuItem = useCallback(
-    (item: NavbarItemType, pin: boolean) => {
-      setPinned(item, pin)
-    },
-    [setPinned]
-  )
   useRepoImportEvents()
 
   return (
     <>
-      <Sidebar.Provider className="min-h-svh">
-        <SidebarView
-          showMoreMenu={showMoreMenu}
-          showSettingMenu={showSettingMenu}
-          handleMoreMenu={handleMoreMenu}
-          handleSettingsMenu={handleSettingsMenu}
-          currentUser={currentUser}
-          handleCustomNav={handleCustomNav}
-          handleLogOut={handleLogOut}
-          recentMenuItems={recentMenu}
-          pinnedMenuItems={pinnedMenu}
-          handleChangePinnedMenuItem={handleChangePinnedMenuItem}
-          handleRemoveRecentMenuItem={handleRemoveRecentMenuItem}
-          useTranslationStore={useTranslationStore}
-        />
-
-        <Sidebar.Inset>
-          <MainContentLayout breadcrumbs={<Breadcrumbs />}>
-            <Outlet />
-          </MainContentLayout>
-
-          <MoreSubmenu showMoreMenu={showMoreMenu} handleMoreMenu={handleMoreMenu} items={moreMenu} />
-          <SettingsMenu
-            showSettingMenu={showSettingMenu}
-            handleSettingsMenu={handleSettingsMenu}
-            items={settingsMenu}
-          />
-          <ManageNavigation
-            pinnedItems={pinnedMenu}
-            recentItems={recentMenu}
-            navbarMenuData={getNavbarMenuData({ t, routes, spaceId: spaceIdPathParam, repoId })}
-            showManageNavigation={showCustomNav}
-            isSubmitting={false}
-            submitted={false}
-            onSave={handleSave}
-            onClose={handleCustomNav}
-          />
-        </Sidebar.Inset>
-      </Sidebar.Provider>
+      <AppSideBar>
+        <MainContentLayout breadcrumbs={<Breadcrumbs />}>
+          <Outlet />
+        </MainContentLayout>
+      </AppSideBar>
       <Toaster />
     </>
   )
@@ -226,9 +80,11 @@ export const AppShellMFE = memo(() => {
 
   return (
     <>
-      <MainContentLayout breadcrumbs={<BreadcrumbsMFE />} className="min-h-screen text-foreground-2">
-        <Outlet />
-      </MainContentLayout>
+      <AppSideBar>
+        <MainContentLayout breadcrumbs={<BreadcrumbsMFE />} className="min-h-screen text-foreground-2">
+          <Outlet />
+        </MainContentLayout>
+      </AppSideBar>
       <Toaster />
     </>
   )
